// src/settings/types.ts
export interface Settings {
  speakRate: number
  piecesListEnabled: boolean
  dividersEnabled: boolean
  customBoardEnabled: boolean
  obfuscationsEnabled: boolean
  parallax: number
  hoverMode: string
  pieceStyle: string
  blur: number
  blackSegments: string
  blackSegmentsTiming: string
  flashModeEnabled: boolean
  flashDuration: number
  flashInterval: number
}